Fix persistence layer bug
XML would reload on each onCreate() instance.
This would make it so that when you minimized the window and opened the app again, it would load the xml file again.
This is due to the fact that onCreate() is ran each time the app process is killed and revived.
To solve this, we added a static boolean flag that makes it so that LoadPersistenceModel only runs once in the entire run cycle of the app.